pyproject.toml
[project]
name = "ipman-cli"
dynamic = ["version"]
description = "Intelligence Package Manager - Agent skill virtual environment manager"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "twisker" },
]
keywords = ["agent", "skill", "virtual-environment", "package-manager", "cli"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Software Distribution",
]

dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.8",
    "mypy>=1.13",
    "types-PyYAML>=6.0",
]
docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
    "mkdocs-click>=0.8",
]

[project.scripts]
ipman = "ipman.cli.main:cli"

[project.urls]
Homepage = "https://github.com/twisker/ipman"
Repository = "https://github.com/twisker/ipman"
Issues = "https://github.com/twisker/ipman/issues"
Documentation = "https://twisker.github.io/ipman"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "VERSION"
pattern = "(?P<version>.+)"

[tool.hatch.build.targets.wheel]
packages = ["src/ipman"]

# ─── ruff ───────────────────────────────────────
[tool.ruff]
target-version = "py310"
line-length = 88
src = ["src", "tests"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["E501"]
"src/ipman/utils/i18n.py" = ["RUF001"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "RUF", # ruff-specific
]

[tool.ruff.lint.isort]
known-first-party = ["ipman"]

# ─── mypy ───────────────────────────────────────
[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

# ─── pytest ─────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
